% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_get_info.R
\name{wb_get_info}
\alias{wb_get_info}
\title{wb_get_info}
\usage{
wb_get_info(
  .fit,
  .timestart,
  .timestop,
  .status,
  .weights = NULL,
  .inits = log(c(6, 100)),
  .method = "Nelder-Mead",
  .control = list(fnscale = -1, .parscale = c(0.1, 0.1), .maxit = 2000),
  .proportional = TRUE,
  .params_greater_one = FALSE
)
}
\description{
Generic function to create HTML tables from lists of results
}
